Revert "Adapt to new change_checker"
This reverts commit 7dcb8c5b97fafab30f1a5e9b47c4e68dd9818cd1.

+++ scripts/elements/electron.rb
@@ -11,7 +11,6 @@ end
 namespace :electron do
   task :clean do
     Electron::TARGETS.each do |path|
-      path = "#{path}/.node_integrity" if File.basename(path) == 'node_modules'
       if File.exist?(path)
         Shell.rm_rf(path)
         Reporter.removed(self, "removed: #{path}", '')
@@ -30,7 +29,7 @@ namespace :electron do
     Shell.chdir(Paths::ELECTRON) do
       Reporter.log 'Installing Electron libraries'
       Shell.sh 'yarn install'
-      Reporter.done('electron', 'installing', '')
+      Reporter.done(self, 'installing', '')
     end
   end
 
@@ -38,18 +37,28 @@ namespace :electron do
     rustcore_dest = "#{Paths::ELECTRON}/node_modules/rustcore"
     Shell.rm_rf(rustcore_dest)
     FileUtils.mkdir_p rustcore_dest
-    # FileUtils.cp_r Dir["#{Paths::TS_BINDINGS}/*"], rustcore_dest
-    FileUtils.cp_r Dir["#{Paths::TS_BINDINGS}/*"].reject{ |f| File.basename(f) == 'node_modules' }, rustcore_dest
+    puts '1...'
+    FileUtils.cp_r Dir.glob("#{Paths::TS_BINDINGS}/*"), rustcore_dest
+    puts '2...'
     Shell.rm_rf("#{rustcore_dest}/native")
-    platform_dest = "#{rustcore_dest}/node_modules/platform"
-    Shell.rm_rf(platform_dest)
-    FileUtils.mkdir_p platform_dest
-    FileUtils.cp_r Dir["#{Paths::PLATFORM}/*"].reject { |f| File.basename(f) == 'node_modules' }, platform_dest
+    puts '3...'
+    Shell.rm_rf("#{rustcore_dest}/node_modules")
+    puts '4...'
+    Shell.chdir(rustcore_dest) do
+      puts '5...'
+      Reporter.log 'Installing rustcore production libraries for electron'
+      Shell.sh 'yarn install --production'
+    end
+    FileUtils.cp_r(Paths::PLATFORM, "#{rustcore_dest}/node_modules")
   end
 
   task copy_client_debug: 'client:build_dev' do
-    path_to_client = "#{Paths::CLIENT_DIST}/debug"
-    FileUtils.rm_f Electron::DIST
+    path_to_client = "#{Paths::CLIENT_DIST}/debug'}"
+    if File.exist?(path_to_client)
+      # TODO: Oli: check if this is sufficient
+      Reporter.skipped('Client', 'client already exist', '')
+      return
+    end
     FileUtils.mkdir_p(Electron::DIST)
     FileUtils.cp_r path_to_client, Electron::DIST
     Reporter.done('Client', "delivery to #{Electron::DIST}", '')
@@ -57,8 +66,12 @@ namespace :electron do
 
   # def self.delivery(dest, prod, replace)
   task copy_client_prod: 'client:build_prod' do
-    path_to_client = "#{Paths::CLIENT_DIST}/release"
-    FileUtils.rm_f Electron::DIST
+    path_to_client = "#{Paths::CLIENT_DIST}/release'}"
+    # TODO: Oli: check if this is sufficient
+    if File.exist?(path_to_client)
+      Reporter.skipped('Client', 'client already exist', '')
+      return
+    end
     FileUtils.mkdir_p(Electron::DIST)
     FileUtils.cp_r path_to_client, Electron::DIST
     Reporter.done('Client', "delivery to #{Electron::DIST}", '')
@@ -67,26 +80,24 @@ namespace :electron do
   task copy_platform: 'platform:build' do
     platform_dest = "#{Electron::NODE_MODULES}/platform"
     Shell.rm_rf(platform_dest)
-    FileUtils.mkdir_p platform_dest
-    FileUtils.cp_r Dir["#{Paths::PLATFORM}/*"].reject { |f| File.basename(f) == 'node_modules' }, platform_dest
+    FileUtils.cp_r(Paths::PLATFORM, Electron::NODE_MODULES)
   end
 
   task check_environment_and_platform: ['environment:check', 'electron:copy_platform']
 
   task do_build: 'updater:build' do
-    changes_to_electron = ChangeChecker.changes?('electron', Paths::ELECTRON)
+    changes_to_electron = ChangeChecker.changes?(Paths::ELECTRON)
     if changes_to_electron
       begin
         Shell.chdir(Paths::ELECTRON) do
           Shell.sh 'yarn run build'
-          ChangeChecker.reset('electron', Paths::ELECTRON, Electron::TARGETS)
+          ChangeChecker.reset(Paths::ELECTRON, TARGETS)
           Reporter.done('electron', 'built', '')
         end
-      rescue StandardError => e
-        puts "An error of type #{e.class} happened, message is #{e.message}"
-        Reporter.failed('electron', 'build', e.message.to_s)
+      rescue StandardError
+        Reporter.failed('electron', 'build', '')
       end
-      FileUtils.cp "#{Paths::ELECTRON}/package.json", Electron::DIST
+      Shell.sh "cp #{Paths::ELECTRON}/package.json #{DIST}/package.json"
     else
       Reporter.skipped('electron', 'build', '')
     end
